backend/requirements.txt
fastapi
uvicorn[standard]
aiohttp
geoip2
python-dotenv
